Edit: Seperate File & Completed of requirements
